requirements.txt
deep_translator==1.11.4
easyocr==1.7.1
matplotlib==3.8.3
numpy==2.0.1
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==10.4.0
